Move db out
